Refactor run SWAN action
