+++ swan-assist/src/main/java/de/fraunhofer/iem/swan/assist/actions/RunSwanAction.java
@@ -7,31 +7,31 @@
 
 package de.fraunhofer.iem.swan.assist.actions;
 
+import com.intellij.ide.util.PropertiesComponent;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
+import com.intellij.openapi.progress.PerformInBackgroundOption;
+import com.intellij.openapi.progress.ProgressManager;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.ui.DialogWrapper;
-import com.intellij.util.messages.MessageBus;
+import de.fraunhofer.iem.swan.assist.analysis.SwanBackgroundTask;
 import de.fraunhofer.iem.swan.assist.comm.SwanNotifier;
 import de.fraunhofer.iem.swan.assist.data.JSONFileLoader;
-import de.fraunhofer.iem.swan.assist.data.TrainingFileManager;
-import de.fraunhofer.iem.swan.assist.ui.dialog.RunAnalysisDialog;
 import de.fraunhofer.iem.swan.assist.util.Constants;
 import de.fraunhofer.iem.swan.data.Method;
 
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Objects;
 import java.util.Properties;
 import java.util.Set;
 
 /**
- * Action opens dialog for user to set parameters for running SWAN. After which thread is created to run SWAN.
+ * Action runs SWAN using the configuration provided in the SettingsDialog. After which thread is created to run SWAN.
  */
-public class RunSwanAnalysisAction extends AnAction {
+public class RunSwanAction extends AnAction {
 
     protected Set<Method> methods = new HashSet<Method>();
 
@@ -44,7 +44,6 @@ public class RunSwanAnalysisAction extends AnAction {
     public void actionPerformed(AnActionEvent anActionEvent) {
 
         final Project project = anActionEvent.getRequiredData(CommonDataKeys.PROJECT);
-        MessageBus messageBus = project.getMessageBus();
 
         Properties config = new Properties();
         InputStream input = null;
@@ -64,30 +63,27 @@ public class RunSwanAnalysisAction extends AnAction {
             }
         }
 
-        //Launch Dialog
-        RunAnalysisDialog dialog = new RunAnalysisDialog(project, true);
-        dialog.show();
-
-        if (dialog.getExitCode() == DialogWrapper.OK_EXIT_CODE) {
-
-            HashMap<String, String> swanParameters = dialog.getParameters();
-
-            if(!swanParameters.get(Constants.CONFIGURATION_FILE).contentEquals(config.getProperty("swan_default_param_value"))){
-                String outputPath = swanParameters.get(Constants.OUTPUT_DIRECTORY) + File.separator + config.getProperty("input_json_suffix");
-                //TrainingFileManager trainingFileManager = new TrainingFileManager(project);
-
-                //if (trainingFileManager.mergeExport(JSONFileLoader.getAllMethods(), outputPath))
-                  //  swanParameters.put(Constants.CONFIGURATION_FILE, outputPath);
-            }
-
-            RunSwanAnalysisImpl processBuilder = new RunSwanAnalysisImpl(project, dialog.getParameters());
-            processBuilder.run();
-
-            SwanNotifier publisher = messageBus.syncPublisher(SwanNotifier.START_SWAN_PROCESS_TOPIC);
-            publisher.launchSwan(null);
+        if (!PropertiesComponent.getInstance(project).isTrueValue(Constants.SWAN_SETTINGS)) {
+            anActionEvent.getActionManager().getAction("SWAN_Assist.SettingsAction").actionPerformed(anActionEvent);
+            //TODO Run SWAN if the tool has been configured
+        } else {
+            runSwan(project);
         }
+    }
 
-        }
+    public void runSwan(Project project) {
+
+        File outputFolder = new File(Objects.requireNonNull(PropertiesComponent.getInstance(project).getValue(Constants.OUTPUT_DIRECTORY)));
+
+        if (!outputFolder.exists())
+            outputFolder.mkdir();
+
+        ProgressManager.getInstance().run(new SwanBackgroundTask(project, "Detecting SRMs", true,
+                PerformInBackgroundOption.ALWAYS_BACKGROUND));
+
+        SwanNotifier publisher = project.getMessageBus().syncPublisher(SwanNotifier.START_SWAN_PROCESS_TOPIC);
+        publisher.launchSwan(null);
+    }
 
     /**
      * Controls whether the action is enabled or disabled
